Switch to E6, E7 and E8 in dimensions 6, 7 and 8.
