+++ src/ChargeFlipPhaser.jl
@@ -1,7 +1,7 @@
 module ChargeFlipPhaser
 import StaticArrays: SMatrix, SVector, MMatrix
 import SparseArrays: sparse, nnz, SparseMatrixCSC
-import LinearAlgebra: norm, transpose, inv, ⋅, Symmetric, tr, mul!, axpy!, det, svd
+import LinearAlgebra: norm, transpose, inv, ⋅, Symmetric, tr, mul!, axpy!, det, svd, qr
 import SpaceGroups: RealOrbit, ComplexOrbit, PhysicalOrbit, ExtinctOrbit, SpaceGroupQuotient, make_orbit
 import FFTW: plan_irfft, plan_rfft, mul!, set_num_threads, irfft
 import Statistics: median
@@ -19,6 +19,7 @@ export WeightedF0, DiffractionData, add_peak!, find_injective_projector,
 
 include("types.jl")
 include("f0_waaskirf.jl")
+include("dense_lattice.jl")
 include("diffractiondata.jl")
 include("windowing.jl")
 include("phaser.jl")
